Move(#12): 메인 페이지에서만 사용되는 모달 컴포넌트들을 main 폴더로 이동
